refactor(archiver): Reorganize files
Removes unneeded `archiver` nested folder, deletes unused `rpc`, moves
helpers under a `modules` subfolder, renames a few internal classes, and
moves a lot of files around within the archiver package.
