+++ yarn-project/archiver/package.json
@@ -4,10 +4,8 @@
   "type": "module",
   "exports": {
     ".": "./dest/index.js",
-    "./data-retrieval": "./dest/archiver/data_retrieval.js",
-    "./epoch": "./dest/archiver/epoch_helpers.js",
     "./test": "./dest/test/index.js",
-    "./config": "./dest/archiver/config.js"
+    "./config": "./dest/config.js"
   },
   "typedocOptions": {
     "entryPoints": [
