+++ yarn-project/archiver/src/errors.ts
