+++ yarn-project/archiver/src/l1/validate_trace.ts
