+++ yarn-project/archiver/src/modules/instrumentation.ts
