+++ yarn-project/archiver/src/modules/validation.ts
