+++ yarn-project/archiver/src/store/block_store.ts
