+++ yarn-project/archiver/src/store/contract_class_store.ts
