+++ yarn-project/archiver/src/structs/inbox_message.ts
