+++ yarn-project/archiver/src/structs/published.ts
